mega_market/requirements.txt
django==2.2.16
djangorestframework==3.12.4
gunicorn==20.0.4
psycopg2-binary==2.8.6
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.1
